requirements.txt
 streamlit
requests
google-generativeai
apscheduler
pytz
python-dotenv